+++ src/tataku/gameplay/game_mode/game_mode_properties.rs
@@ -7,7 +7,7 @@ pub trait GameModeProperties: Send + Sync {
     /// should the cursor be visible (ie, osu yes, taiko/mania no)
     fn show_cursor(&self) -> bool { false }
     /// what should the cursor ripple size be? None for default
-    fn ripple_size(&self) -> Option<f64> { None }
+    fn ripple_size(&self) -> Option<f32> { None }
     
     /// what ms does this map end?
     fn end_time(&self) -> f32;
