+++ src/tataku/gameplay/gameplay_helpers/hit_indicator.rs
@@ -12,9 +12,9 @@ pub trait JudgementIndicator: Send + Sync{
 pub struct BasicJudgementIndicator {
     pub pos: Vector2,
     pub time: f32,
-    pub depth: f64,
+    pub depth: f32,
 
-    pub radius: f64,
+    pub radius: f32,
     pub color: Color,
 
     pub image: Option<Image>,
@@ -24,7 +24,7 @@ pub struct BasicJudgementIndicator {
 impl BasicJudgementIndicator {
     /// pos, depth, radius and color are only if image is none.
     /// if image is some, it assumes the values (pos, depth, size, etc) are already set
-    pub fn new(pos: Vector2, time: f32, depth: f64, radius: f64, color: Color, image: Option<Image>) -> Self {
+    pub fn new(pos: Vector2, time: f32, depth: f32, radius: f32, color: Color, image: Option<Image>) -> Self {
         Self {
             pos,
             time,
