+++ src/tataku/gameplay/modes/mania/mania_notes/mania_hitobject.rs
@@ -8,7 +8,7 @@ pub trait ManiaHitObject: HitObject {
     fn was_hit(&self) -> bool { false }
     fn get_hitsound(&self) -> &Vec<Hitsound>;
 
-    fn set_sv_mult(&mut self, sv: f64);
+    fn set_sv_mult(&mut self, sv: f32);
     fn set_position_function(&mut self, p: Arc<Vec<PositionPoint>>);
     fn playfield_changed(&mut self, playfield: Arc<ManiaPlayfield>);
     fn set_skin_settings(&mut self, settings: Option<Arc<ManiaSkinSettings>>);
